Move application service to domain layer
