+++ admin-site/admin-site-back/src/main/java/eu/adainius/newsfocused/admin/site/back/domain/services/AdminSiteApplicationService.java
@@ -1,4 +1,4 @@
-package eu.adainius.newsfocused.admin.site.back.services;
+package eu.adainius.newsfocused.admin.site.back.domain.services;
 
 import java.util.Optional;
 import java.util.Set;
