+++ admin-site/admin-site-back/src/main/java/eu/adainius/newsfocused/admin/site/back/infrastructure/controller/ApiController.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Controller;
 
 import eu.adainius.newsfocused.admin.site.back.domain.User;
 import eu.adainius.newsfocused.admin.site.back.domain.UserSearchResults;
-import eu.adainius.newsfocused.admin.site.back.services.AdminSiteApplicationService;
+import eu.adainius.newsfocused.admin.site.back.domain.services.AdminSiteApplicationService;
 import io.vavr.control.Either;
 import lombok.AllArgsConstructor;
